requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
dateparser==0.7.2
defusedxml==0.6.0
dj-stripe==2.3.0
Django==2.2.7
django-allauth==0.40.0
django-countries==6.1.2
django-crispy-forms==1.8.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-filter==2.2.0
django-role-permissions==2.2.1
djangorestframework==3.11.0
dnspython==1.16.0
fake-factory==9999.9.9
Faker==2.0.4
gunicorn==20.0.4
idna==2.8
isort==4.3.21
jet-bridge-base==0.6.6
jet-django==0.8.9
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.19
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.2.1
prompt-toolkit==2.0.9
protobuf==3.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2019.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
stripe==2.47.0
text-unidecode==1.3
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
